# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "VentasFix API",
        "description": "API REST del backoffice de VentasFix para la gestión de usuarios, productos y clientes. Pensada para el consumo desde aplicaciones de terceros (por ejemplo, el sistema de gestión Softland). Todos los endpoints protegidos requieren un token Bearer obtenido en /api/login.",
        "contact": {
            "name": "Felipe Jiménez",
            "email": "felipeandresjb99@gmail.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Servidor local de desarrollo"
        }
    ],
    "paths": {
...
        "/api/productos": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Listar todos los productos",
                "operationId": "786d74243fbf30dace82b25a638113b4",
                "responses": {
                    "200": {
                        "description": "Listado de productos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Producto"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Crear un nuevo producto",
                "description": "El precio_de_venta se calcula automáticamente aplicando 19% de IVA sobre el precio_neto. La imagen se envía como archivo (multipart/form-data).",
                "operationId": "fc219fac91d809b8de37c41b5bf9f22a",
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "nombre",
                                    "descripcion_corta",
                                    "descripcion_larga",
                                    "imagen",
                                    "precio_neto",
                                    "stock_actual",
                                    "stock_minimo",
                                    "stock_bajo",
                                    "stock_alto"
                                ],
                                "properties": {
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-001"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Teclado mecánico"
                                    },
                                    "descripcion_corta": {
                                        "type": "string",
                                        "example": "Teclado RGB"
                                    },
                                    "descripcion_larga": {
                                        "type": "string",
                                        "example": "Teclado mecánico retroiluminado."
                                    },
                                    "imagen": {
                                        "description": "Archivo de imagen (jpeg, png, jpg, webp — máx 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "precio_neto": {
                                        "type": "integer",
                                        "example": 10000
                                    },
                                    "stock_actual": {
                                        "type": "integer",
                                        "example": 50
                                    },
                                    "stock_minimo": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "stock_bajo": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "stock_alto": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Producto creado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto creado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/productos/{producto}": {
            "get": {
                "tags": [
                    "Productos"
                ],
                "summary": "Obtener un producto por su ID",
                "operationId": "eeaac99c7b7d7e2ef13b45592b0c3cb4",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datos del producto",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Producto"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Productos"
                ],
                "summary": "Actualizar un producto por su ID",
                "description": "Para actualizar enviando la imagen use POST con multipart/form-data e incluya el campo _method=PUT (Laravel form method spoofing). La imagen es opcional; si se omite se conserva la actual.",
                "operationId": "6021ee99315789e455544f3efe74536c",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "sku",
                                    "nombre",
                                    "descripcion_corta",
                                    "descripcion_larga",
                                    "precio_neto",
                                    "stock_actual",
                                    "stock_minimo",
                                    "stock_bajo",
                                    "stock_alto"
                                ],
                                "properties": {
                                    "_method": {
                                        "type": "string",
                                        "example": "PUT"
                                    },
                                    "sku": {
                                        "type": "string",
                                        "example": "SKU-001"
                                    },
                                    "nombre": {
                                        "type": "string",
                                        "example": "Teclado mecánico"
                                    },
                                    "descripcion_corta": {
                                        "type": "string",
                                        "example": "Teclado RGB"
                                    },
                                    "descripcion_larga": {
                                        "type": "string",
                                        "example": "Teclado mecánico retroiluminado."
                                    },
                                    "imagen": {
                                        "description": "Opcional. Archivo de imagen (jpeg, png, jpg, webp — máx 2MB)",
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "precio_neto": {
                                        "type": "integer",
                                        "example": 12000
                                    },
                                    "stock_actual": {
                                        "type": "integer",
                                        "example": 45
                                    },
                                    "stock_minimo": {
                                        "type": "integer",
                                        "example": 5
                                    },
                                    "stock_bajo": {
                                        "type": "integer",
                                        "example": 10
                                    },
                                    "stock_alto": {
                                        "type": "integer",
                                        "example": 100
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Producto actualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Producto actualizado correctamente."
                                        },
                                        "data": {
                                            "$ref": "#/components/schemas/Producto"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Productos"
                ],
                "summary": "Eliminar un producto por su ID",
                "operationId": "6ac14620c02a5fee41b3e626f8037665",
                "parameters": [
                    {
                        "name": "producto",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Producto eliminado"
                    },
                    "401": {
                        "description": "No autenticado"
                    },
                    "404": {
                        "description": "Producto no encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Producto": {
                "title": "Producto",
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "sku": {
                        "type": "string",
                        "example": "SKU-001"
                    },
                    "nombre": {
                        "type": "string",
                        "example": "Teclado mecánico"
                    },
                    "descripcion_corta": {
                        "type": "string",
                        "example": "Teclado RGB"
                    },
                    "descripcion_larga": {
                        "type": "string",
                        "example": "Teclado mecánico retroiluminado con switches azules."
                    },
                    "imagen": {
                        "type": "string",
                        "example": "productos/abc123.jpg"
                    },
                    "precio_neto": {
                        "type": "integer",
                        "example": 10000
                    },
                    "precio_de_venta": {
                        "description": "Precio con IVA (19%), calculado automáticamente",
                        "type": "integer",
                        "example": 11900
                    },
                    "stock_actual": {
                        "type": "integer",
                        "example": 50
                    },
                    "stock_minimo": {
                        "type": "integer",
                        "example": 5
                    },
                    "stock_bajo": {
                        "type": "integer",
                        "example": 10
                    },
                    "stock_alto": {
                        "type": "integer",
                        "example": 100
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "nullable": true
                    }
                },
                "type": "object"
            },
...
            "ValidationError": {
                "title": "Error de validación",
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "type": "object",
                        "example": {
                            "email": [
                                "El campo email es obligatorio."
                            ]
                        }
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Introduce el token entregado por /api/login como: Bearer {token}",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}